# docs/swagger.yaml
definitions:
  handler.errorResponse:
    properties:
      message:
        type: string
    type: object
  scraping.Category:
    properties:
      id:
        type: string
      name:
        type: string
    type: object
  scraping.Ingredients:
    properties:
      name:
        type: string
      value:
        type: string
    type: object
  scraping.Preview:
    properties:
      author:
        type: string
      comment:
        type: string
      id:
        type: string
      name:
        type: string
      photo:
        type: string
    type: object
  scraping.Recipe:
    properties:
      comment:
        type: string
      id:
        type: string
      ingredients:
        items:
          $ref: '#/definitions/scraping.Ingredients'
        type: array
      name:
        type: string
      photo:
        type: string
      steps:
        items:
          $ref: '#/definitions/scraping.Step'
        type: array
    type: object
  scraping.Step:
    properties:
      comment:
        items:
          type: string
        type: array
      photo:
        type: string
    type: object
host: 195.2.80.162:80
info:
  contact: {}
  description: API Server for povarenokApp
  title: Povarenok :)
  version: "1.0"
paths:
  /recipe/categories:
    get:
      consumes:
      - application/json
      operationId: getCategories
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            type: integer
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handler.errorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/handler.errorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handler.errorResponse'
      tags:
      - Categories
  /recipe/show:
    get:
      consumes:
      - application/json
      operationId: getRecipe
      parameters:
      - description: id
        in: query
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            type: integer
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handler.errorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/handler.errorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handler.errorResponse'
      tags:
      - Recipe
  /recipes:
    get:
      consumes:
      - application/json
      operationId: getPreview
      parameters:
      - description: category
        in: query
        name: category
        type: string
      - description: page
        in: query
        name: page
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            type: integer
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handler.errorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/handler.errorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handler.errorResponse'
      tags:
      - Preview
swagger: "2.0"